created title screen room
i redid obj_load and will continute to work on it later. for now it will be a useless husk of missing objects. i also fixed obj_wall aka invisible walls so now they *should* collide with the player. other than that im just working on getting a title screen in the works.